+++ projects/rocprofiler-compute/src/utils/file_io.py
@@ -19,7 +19,11 @@ from utils.logger import (
     console_warning,
     demarcate,
 )
-from utils.utils_common import canonical_config_arch, normalize_filter_to_str_list
+from utils.utils_common import (
+    _PROFILE_OUTPUT_FORMAT,
+    canonical_config_arch,
+    normalize_filter_to_str_list,
+)
 
 # TODO: use pandas chunksize or dask to read really large csv file
 # from dask import dataframe as dd